adding EnhancedDialog to DeleteContentTypeDialog
