+++ ui/app/src/components/DeleteContentTypeDialog/index.ts
@@ -1,3 +1,7 @@
 export { default } from './DeleteContentTypeDialog';
 
 export * from './DeleteContentTypeDialog';
+
+export * from './DeleteContentTypeDialogContainer';
+
+export * from './utils';
