+++ ui/app/src/components/DeleteContentTypeDialog/utils.ts
@@ -1,0 +1,43 @@
+/*
+ * Copyright (C) 2007-2021 Crafter Software Corporation. All Rights Reserved.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 3 as published by
+ * the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+import ContentType from '../../models/ContentType';
+import { Resource } from '../../models/Resource';
+import { FetchContentTypeUsageResponse } from '../../services/contentTypes';
+import { EnhancedDialogProps } from '../EnhancedDialog';
+import React from 'react';
+import { onSubmittingAndOrPendingChangeProps } from '../../utils/hooks/useEnhancedDialogState';
+
+export interface DeleteContentTypeBaseProps {
+  contentType: ContentType;
+}
+
+export interface DeleteContentTypeDialogProps extends DeleteContentTypeBaseProps, EnhancedDialogProps {
+  onSubmittingAndOrPendingChange(value: onSubmittingAndOrPendingChangeProps): void;
+  onComplete?();
+}
+
+export interface DeleteContentTypeDialogContainerProps
+  extends DeleteContentTypeBaseProps,
+    Pick<DeleteContentTypeDialogProps, 'onComplete' | 'onClose' | 'isSubmitting' | 'onSubmittingAndOrPendingChange'> {}
+
+export interface DeleteContentTypeDialogBodyProps {
+  submitting: boolean;
+  contentType: ContentType;
+  resource: Resource<FetchContentTypeUsageResponse>;
+  password?: string;
+  onCloseButtonClick?(e: React.MouseEvent<HTMLButtonElement, MouseEvent>): void;
+  onSubmit(): void;
+}
