+++ ui/app/src/components/PublishDialog/PublishDialogContentUI.tsx
@@ -18,7 +18,7 @@ import Grid from '@mui/material/Grid';
 import DependencySelection from '../../modules/Content/Dependencies/DependencySelection';
 import PublishDialogForm from './PublishDialogForm';
 import React from 'react';
-import { PublishDialogUIProps } from './PublishDialogUI';
+import { PublishDialogUIProps } from './utils';
 
 export type PublishDialogContentUIProps = Pick<
   PublishDialogUIProps,
