+++ ui/app/src/components/PublishDialog/index.ts
@@ -25,3 +25,5 @@ export * from './PublishDialogUI';
 export * from './PublishDialogContainer';
 
 export * from './PublishDialogForm';
+
+export * from './utils';
